Migrate backend database from MySQL to SQLite
- Replace go-sql-driver/mysql with modernc.org/sqlite (CGO-free)
- Update schema.sql for SQLite syntax (AUTOINCREMENT, remove ON UPDATE)
- Update queries: INSERT OR IGNORE, ON CONFLICT DO UPDATE, CURRENT_TIMESTAMP
- Initialize schema automatically on connect via embedded schema.sql
- Enable WAL mode and foreign keys via PRAGMAs
- Update sqlc.yaml engine to sqlite with type overrides
- Adapt type changes: Rating int64, Price NullInt64, AvgRating NullFloat64
- Remove MySQL service from docker-compose.dev.yml
- Default DSN changed to beanmemo.db
- Add *.db to .gitignore

+++ backend/internal/database/sqlcgen/models.go
@@ -13,11 +13,11 @@ type Record struct {
 	ID           int64          `json:"id"`
 	UserID       int64          `json:"user_id"`
 	Name         string         `json:"name"`
-	Rating       int8           `json:"rating"`
+	Rating       int64          `json:"rating"`
 	Origin       sql.NullString `json:"origin"`
 	RoastLevel   sql.NullString `json:"roast_level"`
 	Shop         sql.NullString `json:"shop"`
-	Price        sql.NullInt32  `json:"price"`
+	Price        sql.NullInt64  `json:"price"`
 	PurchasedAt  sql.NullTime   `json:"purchased_at"`
 	TastingNote  sql.NullString `json:"tasting_note"`
 	BrewMethod   sql.NullString `json:"brew_method"`
